+++ modules/mesh-task/main.tf
@@ -15,6 +15,12 @@ locals {
 
   consul_binary_volume_name = "consul_binary"
 
+  // Optionally, users can override the application container's entrypoint.
+  app_entrypoint = var.application_shutdown_delay_seconds == null ? null : [
+    "/consul/consul-ecs", "app-entrypoint", "-shutdown-delay", "${var.application_shutdown_delay_seconds}s",
+  ]
+  app_mountpoints = var.application_shutdown_delay_seconds == null ? [] : [local.consul_data_mount]
+
   // container_defs_with_depends_on is the app's container definitions with their dependsOn keys
   // modified to add in dependencies on consul-ecs-mesh-init and sidecar-proxy.
   // We add these dependencies in so that the app containers don't start until the proxy
@@ -37,6 +43,16 @@ locals {
               }
             ]
         ))
+      },
+      {
+        // Use the def.entryPoint, if defined. Else, use the app_entrypoint, which is null by default.
+        entryPoint = lookup(def, "entryPoint", local.app_entrypoint)
+        mountPoints = flatten(
+          concat(
+            lookup(def, "mountPoints", []),
+            local.app_mountpoints,
+          )
+        )
       }
     )
   ]
@@ -258,7 +274,7 @@ resource "aws_ecs_task_definition" "this" {
             image            = var.envoy_image
             essential        = false
             logConfiguration = var.log_configuration
-            entrypoint       = ["/consul/consul-ecs", "envoy-entrypoint"]
+            entryPoint       = ["/consul/consul-ecs", "envoy-entrypoint"]
             command          = ["envoy", "--config-path", "/consul/envoy-bootstrap.json"]
             portMappings = [
               {