+++ CHANGELOG.md
@@ -7,6 +7,10 @@ FEATURES
 * modules/mesh-task: Add `task_role_arn` and `execution_role_arn` input variables
   which specify the task and execution role to include in the task definition.
   [[GH-71](https://github.com/hashicorp/terraform-aws-consul-ecs/pull/71)]
+* modules/mesh-task: Add `application_shutdown_delay_seconds` variable to
+  delay application shutdown. This allows time for incoming traffic to drain
+  off for better graceful shutdown.
+  [[GH-67](https://github.com/hashicorp/terraform-aws-consul-ecs/pull/67)]
 
 ## 0.2.0 (Nov 16, 2021)